Fixes for HA fencing plugin
* Document a w/a for a parrallel deploy issue
* Add known issues section to docs
* Add missing versioning section to docs
* Deploy the plugin only at controller nodes

Related blueprint fencing-in-puppet-manifests
Related-bug: #1411603

+++ ha_fencing/deployment_scripts/puppet/manifests/site.pp
@@ -1,21 +1,26 @@
 $fuel_settings = parseyaml($astute_settings_yaml)
 
-# Fetch fencing policy and settings
-$fence_policy = $::fuel_settings['ha_fencing']['fence_policy']
-$fencing_enabled  = $fence_policy ? { 'disabled'=>false, 'reboot'=>true, 'poweroff'=>true, default=>false }
+$primary_controller = $::fuel_settings['role'] ? { 'primary-controller'=>true, default=>false }
+$is_controller      = $::fuel_settings['role'] ? { 'controller'=>true, default=>false }
 
-if $fencing_enabled {
-  $fencing_settings = parseyaml($fencing_settings_yaml)
-  $fence_primitives = $::fencing_settings['fence_primitives']
-  $fence_topology   = $::fencing_settings['fence_topology']
+if ($is_controller or $primary_controller) {
+  # Fetch fencing policy and settings
+  $fence_policy = $::fuel_settings['ha_fencing']['fence_policy']
+  $fencing_enabled  = $fence_policy ? { 'disabled'=>false, 'reboot'=>true, 'poweroff'=>true, default=>false }
 
-  $nodes_hash = $::fuel_settings['nodes']
-  $controllers = concat(filter_nodes($nodes_hash,'role','primary-controller'), filter_nodes($nodes_hash,'role','controller'))
+  if $fencing_enabled {
+    $fencing_settings = parseyaml($fencing_settings_yaml)
+    $fence_primitives = $::fencing_settings['fence_primitives']
+    $fence_topology   = $::fencing_settings['fence_topology']
 
-  include stdlib
-  class { '::pcs_fencing::fencing_primitives':
-    fence_primitives => $fence_primitives,
-    fence_topology   => $fence_topology,
-    nodes            => $controllers,
+    $nodes_hash = $::fuel_settings['nodes']
+    $controllers = concat(filter_nodes($nodes_hash,'role','primary-controller'), filter_nodes($nodes_hash,'role','controller'))
+
+    include stdlib
+    class { '::pcs_fencing::fencing_primitives':
+      fence_primitives => $fence_primitives,
+      fence_topology   => $fence_topology,
+      nodes            => $controllers,
+    }
   }
 }
